Section 10 like the acctual one
Apparently i messed up a lot and like a lot of things got messed up because of it. Anyways if you are reading this just ignore the numbers on the other commits. I fixed a bunch of stuff i messed up whenever i was copying and pasting it from visual studio.  bunch of stuff also got cut off. This commit fixes everything